refactor scaling columns
