+++ packages/frontend/src/components/table/props/getScalingTableColumns.tsx
@@ -1,0 +1,235 @@
+import React from 'react'
+import { ScalingRiskViewEntry } from '../../../pages/scaling-risk/view/ScalingRiskView'
+import { ScalingTvlViewEntry } from '../../../pages/scaling-tvl/view/ScalingTvlView'
+import { IndexCell } from '../IndexCell'
+import { MaturityCell } from '../MaturityCell'
+import { NumberCell } from '../NumberCell'
+import { ProjectCell } from '../ProjectCell'
+import { RiskCell } from '../RiskCell'
+import { RosetteCell } from '../RosetteCell'
+import { ColumnConfig } from '../TableView'
+import { TechnologyCell } from '../TechnologyCell'
+
+export function getActiveScalingTvlColumns(maturityEnabled: boolean) {
+  const columns: ColumnConfig<ScalingTvlViewEntry>[] = [
+    {
+      name: '#',
+      alignCenter: true,
+      minimalWidth: true,
+      headClassName: 'md:pl-4',
+      getValue: (_, index) => <IndexCell index={index} className="md:pl-4" />,
+    },
+    {
+      name: 'Name',
+      headClassName: 'pl-8',
+      getValue: (project) => <ProjectCell type="layer2" project={project} />,
+    },
+    {
+      name: 'Risks',
+      tooltip: 'Risks associated with this project.',
+      minimalWidth: true,
+      alignCenter: true,
+      getValue: (project) => <RosetteCell riskValues={project.riskValues} />,
+    },
+    {
+      name: 'Technology',
+      tooltip:
+        'Type of this Layer 2. Determines data availability and proof system used.',
+      shortName: 'Tech',
+      getValue: (project) => (
+        <TechnologyCell provider={project.provider}>
+          {project.technology}
+        </TechnologyCell>
+      ),
+    },
+    ...(maturityEnabled
+      ? [
+          {
+            name: 'Maturity',
+            tooltip: 'Maturity of this Layer 2 based on its features.',
+            alignCenter: true as const,
+            getValue: (project: ScalingTvlViewEntry) => (
+              <MaturityCell item={project.maturityEntry} />
+            ),
+          },
+        ]
+      : []),
+    {
+      name: 'Purpose',
+      tooltip: 'Functionality supported by this Layer 2.',
+      getValue: (project) => project.purpose,
+    },
+    {
+      name: 'TVL',
+      tooltip: 'Total value locked in escrow contracts on Ethereum.',
+      alignRight: true,
+      noPaddingRight: true,
+      headClassName: '-translate-x-[72px]',
+      getValue: (project) => (
+        <>
+          <NumberCell className="font-bold">{project.tvl}</NumberCell>
+          <NumberCell signed className="ml-1 w-[72px] !text-base font-medium ">
+            {project.sevenDayChange}
+          </NumberCell>
+        </>
+      ),
+    },
+    {
+      name: 'Mkt share',
+      tooltip: 'Share of the sum of total value locked of all projects.',
+      alignRight: true,
+      minimalWidth: true,
+      headClassName: 'pr-4',
+      getValue: (project) =>
+        project.tvlBreakdown && (
+          <NumberCell className="pr-4">{project.marketShare}</NumberCell>
+        ),
+    },
+  ]
+
+  return columns
+}
+
+export function getUpcomingScalingTvlColumns() {
+  const columns: ColumnConfig<ScalingTvlViewEntry>[] = [
+    {
+      name: '#',
+      alignCenter: true,
+      minimalWidth: true,
+      headClassName: 'md:pl-4',
+      getValue: (_, index) => <IndexCell index={index} className="md:pl-4" />,
+    },
+    {
+      name: 'Name',
+      headClassName: 'pl-8',
+      getValue: (project) => <ProjectCell type="layer2" project={project} />,
+    },
+    {
+      name: 'Technology',
+      tooltip:
+        'Type of this Layer 2. Determines data availability and proof system used.',
+      shortName: 'Tech',
+      getValue: (project) => (
+        <TechnologyCell provider={project.provider}>
+          {project.technology}
+        </TechnologyCell>
+      ),
+    },
+    {
+      name: 'Purpose',
+      tooltip: 'Functionality supported by this Layer 2.',
+      getValue: (project) => project.purpose,
+    },
+  ]
+
+  return columns
+}
+
+export function getArchivedScalingTvlColumns() {
+  const columns: ColumnConfig<ScalingTvlViewEntry>[] = [
+    {
+      name: '#',
+      alignCenter: true,
+      minimalWidth: true,
+      headClassName: 'md:pl-4',
+      getValue: (_, index) => <IndexCell index={index} className="md:pl-4" />,
+    },
+    {
+      name: 'Name',
+      headClassName: 'pl-8',
+      getValue: (project) => <ProjectCell type="layer2" project={project} />,
+    },
+    {
+      name: 'Risks',
+      tooltip: 'Risks associated with this project.',
+      minimalWidth: true,
+      alignCenter: true,
+      getValue: (project) => <RosetteCell riskValues={project.riskValues} />,
+    },
+    {
+      name: 'Technology',
+      tooltip:
+        'Type of this Layer 2. Determines data availability and proof system used.',
+      shortName: 'Tech',
+      getValue: (project) => (
+        <TechnologyCell provider={project.provider}>
+          {project.technology}
+        </TechnologyCell>
+      ),
+    },
+    {
+      name: 'Purpose',
+      tooltip: 'Functionality supported by this Layer 2.',
+      getValue: (project) => project.purpose,
+    },
+    {
+      name: 'TVL',
+      tooltip: 'Total value locked in escrow contracts on Ethereum.',
+      alignRight: true,
+      noPaddingRight: true,
+      headClassName: '-translate-x-[72px]',
+      getValue: (project) => (
+        <>
+          <NumberCell className="font-bold">{project.tvl}</NumberCell>
+          {!project.isArchived ? (
+            <NumberCell
+              signed
+              className="ml-1 w-[72px] !text-base font-medium "
+            >
+              {project.sevenDayChange}
+            </NumberCell>
+          ) : (
+            <span className="w-[72px]" />
+          )}
+        </>
+      ),
+    },
+  ]
+
+  return columns
+}
+
+export function getScalingRiskColumns() {
+  const columns: ColumnConfig<ScalingRiskViewEntry>[] = [
+    {
+      name: '#',
+      alignCenter: true,
+      minimalWidth: true,
+      headClassName: 'md:pl-4',
+      getValue: (_, index) => <IndexCell index={index} className="md:pl-4" />,
+    },
+    {
+      name: 'Name',
+      headClassName: 'pl-8',
+      getValue: (project) => <ProjectCell type="layer2" project={project} />,
+    },
+    {
+      name: 'State validation',
+      tooltip: 'How is the validity of the system state checked?',
+      getValue: (project) => <RiskCell item={project.stateValidation} />,
+    },
+    {
+      name: 'Data availability',
+      tooltip: 'Is the data needed to reconstruct the state available?',
+      getValue: (project) => <RiskCell item={project.dataAvailability} />,
+    },
+    {
+      name: 'Upgradeability',
+      tooltip: 'Are the Ethereum contracts upgradeable?',
+      getValue: (project) => <RiskCell item={project.upgradeability} />,
+    },
+    {
+      name: 'Sequencer failure',
+      tooltip:
+        "Sequencer is an entity responsible for constructing L2 blocks and deciding on the ordering of user's L2 transactions. What happens if it is offline or censors individual user?",
+      getValue: (project) => <RiskCell item={project.sequencerFailure} />,
+    },
+    {
+      name: 'Validator failure',
+      tooltip:
+        'Validator is an entity responsible for submitting L2 state to Ethereum (optionally, along with the zkProof). What happens if it is offline?',
+      getValue: (project) => <RiskCell item={project.validatorFailure} />,
+    },
+  ]
+  return columns
+}
